refactor(telegram): drop the turn WAL — in-memory per-session queue
The turn store's file-backed WAL (recovery of `queued` turns, `started`
loud-drop, redelivery tombstones, shape validation) was a partial durability
layer bolted onto the real feature: the per-session FIFO serializer that shipped
in-memory first and never depended on it. Telegram already redelivers
never-ACKed updates on its own; the WAL's only added window was an
ACKed-but-unfinished turn lost to a crash — a window a single-process bot
accepts (the asker re-asks), and whose proper fix is durable execution at a
K-axis backend (an external queue with distributed-locking recovery), not a
hand-rolled per-process log in the channel.

Collapsing the WAL cascades: no recovery ⇒ no replay ⇒ the tombstone/suppressed
dedup (which only existed to stop a WAL-replay from double-running a Telegram
redelivery) is unnecessary too. `turn-store.ts` (186 lines) becomes
`turn-queue.ts` (~40) — the in-memory FIFO chains it always had. telegram.ts
sheds isPendingTurn, the suppressed gate, and the persisted-record fields; the
⏳ queued-notice id is now a direct mutation on the shared record (gated by the
same dequeue await), not a store-owned patch.

The un-summoned context buffer stays durable — its case is stronger (those
messages are never redelivered as a summon). This also makes turn durability
consistent with start's existing no-graceful-drain on SIGTERM.

Docs (core.md §9.2, telegram.md, module headers, AGENTS.md) and CHANGELOG record
the accepted loss and the upgrade note (a prior queue.json is no longer read).

Net -317 lines. Tests: WAL recovery matrix removed; FIFO/concurrency/queue-notice
kept; turn-queue.test.ts covers the in-memory queue incl. the runner-rejection
backstop. typecheck + 326 tests green.
